adk_agent_portal26/requirements.txt
google-adk
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-google-genai
python-dotenv
